DDF-4043 Updated tested and created utility class instead of inheritance
